% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mgr.R
\name{predict_mgr}
\alias{predict_mgr}
\title{Predict Mycorrhizal Growth Response}
\usage{
predict_mgr(LDMC, SLA, LCC, LNC, RD, SRL, RTD, SRA, RCC, RNC, precipitation)
}
\arguments{
\item{LDMC}{Leaf Dry Matter Content}

\item{SLA}{Specific Leaf Area}

\item{LCC}{Leaf Carbon Concentration}

\item{LNC}{Leaf Nitrogen Concentration}

\item{RD}{Root Diameter}

\item{SRL}{Specific Root Length}

\item{RTD}{Root Tissue Density}

\item{SRA}{Specific Root Area}

\item{RCC}{Root Carbon Concentration}

\item{RNC}{Root Nitrogen Concentration}

\item{precipitation}{Precipitation}
}
\value{
Predicted Mycorrhizal Growth Response
}
\description{
Predict MGR from plant functional traits.
}
